refactor(frontend): split ProteinStructure3D into sub-components (#133)
ProteinStructure3D.vue was 1,132 lines (>2x the 500-line guideline) with
multiple responsibilities tangled together. Split it into a slim smart
container plus four single-responsibility sub-components, preserving the
exact public contract and all behavior.

Structure (parent 1132 -> 416 lines):
- protein-structure/StructureViewer.vue   — owns the NGL canvas + ALL
  imperative NGL logic (stage/representations/distance calc, markRaw,
  /2h8r.cif load, useLegacyLights suppression). Single NGL owner.
- protein-structure/StructureControls.vue — representation toggle, DNA /
  domain switches, reset + distance-line buttons (presentational).
- protein-structure/VariantPanel.vue      — filter/sort selects + variant
  list + empty state (all-variants mode, presentational).
- protein-structure/DistanceStatsCard.vue — distance-to-DNA alert + legend
  (presentational).

Wiring is pure props-down / emits-up (house pattern); no provide/inject,
no store. The parent keeps the EXACT public API (props variants /
currentVariantId / showAllVariants, emits variant-clicked), so the two
call sites (PageVariant.vue, Home.vue) and the existing characterization
unit test are untouched. Per-variant DNA distances are computed once by
the viewer and emitted up so the panel/stats stay presentational.

The issue's proposed StructureTooltip.vue was dropped: there is no hover
tooltip in the current code, so it would be net-new feature work rather
than extraction.

Tests: existing ProteinStructure3D.spec.js stays green; added NGL-free
unit specs for the three presentational children (479 tests pass).
Runtime-verified with Playwright in both modes (single-variant detail +
all-variants Home): structure renders, variant panel + distances
populate, controls/representation/domain-colors/reset/distance-line all
work, DistanceStatsCard shows real NGL-computed distances, zero console
errors.

Closes #133
